+++ src/watcher.js
@@ -9,6 +9,7 @@ const { log, clog, error } = require('./utils/logger')
 
 const execWatcher = async (chokidarOpts, destination, fileID, clean, execAfter) => {
   const watcher = chokidar.watch(chokidarOpts.paths, chokidarOpts)
+  let timer = null
 
   const execAfterAction = async () => {
     // clean directory before
@@ -44,9 +45,12 @@ const execWatcher = async (chokidarOpts, destination, fileID, clean, execAfter) 
   // Event listeners
   watcher
     .on('unlink', (path) => log(`removed: ${path}`, 'red'))
-    .on('change', (path) => {
+    .on('change', async (path) => {
       log(`changed: ${path}`, 'yellow')
-      execAfterAction()
+
+      if (timer) clearTimeout(timer)
+      // execute script with timeout, otherwise it is runned for each file change
+      timer = setTimeout(execAfterAction, 600)
     })
     .on('ready', () => {
       execAfterAction()
@@ -64,6 +68,7 @@ const execWatcher = async (chokidarOpts, destination, fileID, clean, execAfter) 
   const { platform, fileID, dest, clean, chokidarOpts, execAfter } = initialize('watcher')
 
   try {
+    // if user set the destination folder manually
     let destination = dest
 
     if (destination) log('Argument --dest provided manually, skipping folder search')