+++ CHANGELOG.md
@@ -1,8 +1,42 @@
-# 1.0.0 (2021-08-26)
+# 1.0.0 (2021-08-27)
 
 
 ### Bug Fixes
 
+* added error message on clean script ([95fef51](https://github.com/PitcherAG/pitcher-watcher/commit/95fef516221a1e75c15c08489b3c4a430889edea))
+* error message for copy script fixed ([c699944](https://github.com/PitcherAG/pitcher-watcher/commit/c699944ce1e5e5ca8e39605e33ae2d249fdbf549))
+* ignoring dot files ([56bd9e1](https://github.com/PitcherAG/pitcher-watcher/commit/56bd9e1d0b680e594bebce3249324578ee8b26ae))
+* issues with cleaning directory ([c1a1dca](https://github.com/PitcherAG/pitcher-watcher/commit/c1a1dca74de2b883d08a483530d1995725391eee))
+* parse clean arg properly ([8b06ab7](https://github.com/PitcherAG/pitcher-watcher/commit/8b06ab737306dbc17b70a1106db04415f03e114a))
+* promisified cleanDirectory ([8368d67](https://github.com/PitcherAG/pitcher-watcher/commit/8368d67f31c708885ce1ac73c813ec0e1d26d544))
+* testing tag ([b715f38](https://github.com/PitcherAG/pitcher-watcher/commit/b715f389f9c46fe89394fdbb290f1b7fe7652a4b))
+
+
+### Features
+
+* add help command ([e7cc819](https://github.com/PitcherAG/pitcher-watcher/commit/e7cc819801e66e8470ee4f00c87f8d5f27412dd1))
+* add timer to watcher after action ([ec57f70](https://github.com/PitcherAG/pitcher-watcher/commit/ec57f700a704887684001284c087f12da64790b9))
+* add windows folder finder ([03ac6a1](https://github.com/PitcherAG/pitcher-watcher/commit/03ac6a11e335653ee6e1aaab8a0185bc6fb4b1e1))
+* app initializer & validator ([ad44af9](https://github.com/PitcherAG/pitcher-watcher/commit/ad44af92437896d60bd5d92da1506cd655e606d7))
+* initial chokidar args ([e19fdda](https://github.com/PitcherAG/pitcher-watcher/commit/e19fddade31e8571d9fcf5e1a4eae19a9d120a83))
+* logger added ([459f15a](https://github.com/PitcherAG/pitcher-watcher/commit/459f15a1969a94a72d0eccd01cfdfa087c78a1c6))
+* merge vueArgs with default script args ([c9f5e06](https://github.com/PitcherAG/pitcher-watcher/commit/c9f5e0667a0e9eae18addffca398ad5a4fcd0c2d))
+* own clean logic ([c84611b](https://github.com/PitcherAG/pitcher-watcher/commit/c84611b1677a9458c57f2a6c1de3d150c31c7f1d))
+* pass vue args to the cli ([57dc03d](https://github.com/PitcherAG/pitcher-watcher/commit/57dc03d182d51ec015733220d1262f7b391ae710))
+* promisify copying file logic for watcher ([bfaff00](https://github.com/PitcherAG/pitcher-watcher/commit/bfaff003e956315127a4370285d1426b5f0e3679))
+* split vue & watcher logic ([ec53b10](https://github.com/PitcherAG/pitcher-watcher/commit/ec53b102147ff583c12375bcfe13468aa0cf3039))
+* support execAfter option ([9ba9ae0](https://github.com/PitcherAG/pitcher-watcher/commit/9ba9ae0cae644379ce7962680d59829dda249977))
+* verbose messages ([5e35c9c](https://github.com/PitcherAG/pitcher-watcher/commit/5e35c9c39d8e6247461d13a655e21bb5f8be974b))
+* watcher ready ([b7b1212](https://github.com/PitcherAG/pitcher-watcher/commit/b7b12128b6b89cfe4439c98ea76e42667a2dde9e))
+* windows logging verbose ([0185cf2](https://github.com/PitcherAG/pitcher-watcher/commit/0185cf239b367f967a386625d236279b6f7520a1))
+
+# 1.0.0 (2021-08-27)
+
+
+### Bug Fixes
+
+* added error message on clean script ([95fef51](https://github.com/PitcherAG/pitcher-watcher/commit/95fef516221a1e75c15c08489b3c4a430889edea))
+* error message for copy script fixed ([c699944](https://github.com/PitcherAG/pitcher-watcher/commit/c699944ce1e5e5ca8e39605e33ae2d249fdbf549))
 * ignoring dot files ([56bd9e1](https://github.com/PitcherAG/pitcher-watcher/commit/56bd9e1d0b680e594bebce3249324578ee8b26ae))
 * issues with cleaning directory ([c1a1dca](https://github.com/PitcherAG/pitcher-watcher/commit/c1a1dca74de2b883d08a483530d1995725391eee))
 * parse clean arg properly ([8b06ab7](https://github.com/PitcherAG/pitcher-watcher/commit/8b06ab737306dbc17b70a1106db04415f03e114a))
@@ -12,6 +46,7 @@
 ### Features
 
 * add help command ([e7cc819](https://github.com/PitcherAG/pitcher-watcher/commit/e7cc819801e66e8470ee4f00c87f8d5f27412dd1))
+* add timer to watcher after action ([ec57f70](https://github.com/PitcherAG/pitcher-watcher/commit/ec57f700a704887684001284c087f12da64790b9))
 * add windows folder finder ([03ac6a1](https://github.com/PitcherAG/pitcher-watcher/commit/03ac6a11e335653ee6e1aaab8a0185bc6fb4b1e1))
 * app initializer & validator ([ad44af9](https://github.com/PitcherAG/pitcher-watcher/commit/ad44af92437896d60bd5d92da1506cd655e606d7))
 * initial chokidar args ([e19fdda](https://github.com/PitcherAG/pitcher-watcher/commit/e19fddade31e8571d9fcf5e1a4eae19a9d120a83))
@@ -19,8 +54,45 @@
 * merge vueArgs with default script args ([c9f5e06](https://github.com/PitcherAG/pitcher-watcher/commit/c9f5e0667a0e9eae18addffca398ad5a4fcd0c2d))
 * own clean logic ([c84611b](https://github.com/PitcherAG/pitcher-watcher/commit/c84611b1677a9458c57f2a6c1de3d150c31c7f1d))
 * pass vue args to the cli ([57dc03d](https://github.com/PitcherAG/pitcher-watcher/commit/57dc03d182d51ec015733220d1262f7b391ae710))
+* promisify copying file logic for watcher ([bfaff00](https://github.com/PitcherAG/pitcher-watcher/commit/bfaff003e956315127a4370285d1426b5f0e3679))
 * split vue & watcher logic ([ec53b10](https://github.com/PitcherAG/pitcher-watcher/commit/ec53b102147ff583c12375bcfe13468aa0cf3039))
 * support execAfter option ([9ba9ae0](https://github.com/PitcherAG/pitcher-watcher/commit/9ba9ae0cae644379ce7962680d59829dda249977))
 * verbose messages ([5e35c9c](https://github.com/PitcherAG/pitcher-watcher/commit/5e35c9c39d8e6247461d13a655e21bb5f8be974b))
 * watcher ready ([b7b1212](https://github.com/PitcherAG/pitcher-watcher/commit/b7b12128b6b89cfe4439c98ea76e42667a2dde9e))
 * windows logging verbose ([0185cf2](https://github.com/PitcherAG/pitcher-watcher/commit/0185cf239b367f967a386625d236279b6f7520a1))
+
+# 1.0.0 (2021-08-27)
+
+
+### Bug Fixes
+
+* added error message on clean script ([95fef51](https://github.com/PitcherAG/pitcher-watcher/commit/95fef516221a1e75c15c08489b3c4a430889edea))
+* error message for copy script fixed ([c699944](https://github.com/PitcherAG/pitcher-watcher/commit/c699944ce1e5e5ca8e39605e33ae2d249fdbf549))
+* ignoring dot files ([56bd9e1](https://github.com/PitcherAG/pitcher-watcher/commit/56bd9e1d0b680e594bebce3249324578ee8b26ae))
+* issues with cleaning directory ([c1a1dca](https://github.com/PitcherAG/pitcher-watcher/commit/c1a1dca74de2b883d08a483530d1995725391eee))
+* parse clean arg properly ([8b06ab7](https://github.com/PitcherAG/pitcher-watcher/commit/8b06ab737306dbc17b70a1106db04415f03e114a))
+* promisified cleanDirectory ([8368d67](https://github.com/PitcherAG/pitcher-watcher/commit/8368d67f31c708885ce1ac73c813ec0e1d26d544))
+
+
+### Features
+
+* add help command ([e7cc819](https://github.com/PitcherAG/pitcher-watcher/commit/e7cc819801e66e8470ee4f00c87f8d5f27412dd1))
+* add timer to watcher after action ([ec57f70](https://github.com/PitcherAG/pitcher-watcher/commit/ec57f700a704887684001284c087f12da64790b9))
+* add windows folder finder ([03ac6a1](https://github.com/PitcherAG/pitcher-watcher/commit/03ac6a11e335653ee6e1aaab8a0185bc6fb4b1e1))
+* app initializer & validator ([ad44af9](https://github.com/PitcherAG/pitcher-watcher/commit/ad44af92437896d60bd5d92da1506cd655e606d7))
+* initial chokidar args ([e19fdda](https://github.com/PitcherAG/pitcher-watcher/commit/e19fddade31e8571d9fcf5e1a4eae19a9d120a83))
+* logger added ([459f15a](https://github.com/PitcherAG/pitcher-watcher/commit/459f15a1969a94a72d0eccd01cfdfa087c78a1c6))
+* merge vueArgs with default script args ([c9f5e06](https://github.com/PitcherAG/pitcher-watcher/commit/c9f5e0667a0e9eae18addffca398ad5a4fcd0c2d))
+* own clean logic ([c84611b](https://github.com/PitcherAG/pitcher-watcher/commit/c84611b1677a9458c57f2a6c1de3d150c31c7f1d))
+* pass vue args to the cli ([57dc03d](https://github.com/PitcherAG/pitcher-watcher/commit/57dc03d182d51ec015733220d1262f7b391ae710))
+* promisify copying file logic for watcher ([bfaff00](https://github.com/PitcherAG/pitcher-watcher/commit/bfaff003e956315127a4370285d1426b5f0e3679))
+* split vue & watcher logic ([ec53b10](https://github.com/PitcherAG/pitcher-watcher/commit/ec53b102147ff583c12375bcfe13468aa0cf3039))
+* support execAfter option ([9ba9ae0](https://github.com/PitcherAG/pitcher-watcher/commit/9ba9ae0cae644379ce7962680d59829dda249977))
+* verbose messages ([5e35c9c](https://github.com/PitcherAG/pitcher-watcher/commit/5e35c9c39d8e6247461d13a655e21bb5f8be974b))
+* watcher ready ([b7b1212](https://github.com/PitcherAG/pitcher-watcher/commit/b7b12128b6b89cfe4439c98ea76e42667a2dde9e))
+* windows logging verbose ([0185cf2](https://github.com/PitcherAG/pitcher-watcher/commit/0185cf239b367f967a386625d236279b6f7520a1))
+
+
+### Reverts
+
+* version 1.1.0 ([4d2c21b](https://github.com/PitcherAG/pitcher-watcher/commit/4d2c21b1a8fc17a2d4bc387973736e50de4fab79))